Reuse dimension_type loading crud function
