+++ breadbox/breadbox/api/temp.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Body, Depends
 
 from breadbox.api.dependencies import get_db_with_user
 from breadbox.config import Settings, get_settings
-from breadbox.crud import dataset as dataset_crud
+from breadbox.crud import types as types_crud
 from breadbox.crud import slice as slice_crud
 from breadbox.db.session import SessionWithUser
 from breadbox.schemas.context import (
@@ -43,9 +43,7 @@ def evaluate_context(
     context_evaluator = ContextEvaluator(context.dict(), slice_loader_function)
 
     # Load all dimension labels and ids
-    all_labels_by_id = dataset_crud.get_dimension_labels_by_id(
-        db, context.dimension_type
-    )
+    all_labels_by_id = types_crud.get_dimension_labels_by_id(db, context.dimension_type)
 
     # Evaluate each against the context
     matching_ids = []
